NO-TICKET: execution-engine: break up execution.rs
